% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_greys.R
\name{add_greys}
\alias{add_greys}
\title{Add greys}
\usage{
add_greys(pal = NULL, grey_levels)
}
\arguments{
\item{pal}{A palette, defaults to \code{NULL}.}

\item{grey_levels}{A vector of grey levels between 0-100.}
}
\description{
Add shades of grey to a palette
}
\examples{
add_greys (grey_levels = c(50,70))
add_greys ( reprex_palette(), grey_levels = c(50,70))
}
\seealso{
Other color functions: 
\code{\link{reprex_palette}()}
}
\concept{color functions}
