% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_colors.R
\name{create_shades}
\alias{create_shades}
\title{Create shades}
\usage{
create_shades(color, palette_length = 11, used_colors = c(11, 8))
}
\arguments{
\item{color}{The color for which you need shades.}

\item{palette_length}{Number of colors to generate.}

\item{used_colors}{Integer indices selecting which colors to return
from the generated palette. Defaults to all colors.
Values must be between 1 and palette_length.}
}
\description{
Create a discrete color palette of the shades of a single color (and white.)
}
\details{
This is a wrapper around
\code{\link{create_palette}} with the parameters
set to \code{color1=color} and \code{color2='white'}.
}
\examples{
create_shades(color = c(violet = "#4E115A"))
}
