% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_european_countries.R
\name{group_european_countries}
\alias{group_european_countries}
\title{Group European Countries}
\usage{
group_european_countries()
}
\value{
A tibble with geo codes and suggested groups.
}
\description{
Group European Countries
}
\examples{
group_european_countries()
}
