% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_numbers.R
\name{pad_numbers}
\alias{pad_numbers}
\title{Pad numbers with leading zeros}
\usage{
pad_numbers(numbers)
}
\arguments{
\item{numbers}{A vector of numbers.}
}
\value{
A vector of strings, padded with leading zeros for equal-length
numbers.
}
\description{
Pad numbers with leading zeros
}
\examples{
pad_numbers(c(1, 12, 132))
}
\keyword{export}
