% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_eu_countries.R
\name{palette_eu_countries}
\alias{palette_eu_countries}
\title{Return a color for each EU country from the SR palette.}
\usage{
palette_eu_countries()
}
\value{
A named vector of color hex strings for each country code.
}
\description{
Return a color for each EU country from the SR palette.
}
\examples{
palette_eu_countries()
}
