% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_europe_countries.R
\name{palette_europe_countries}
\alias{palette_europe_countries}
\title{Return a color for each European country from the SR palette.}
\usage{
palette_europe_countries(europe_countries_4 = FALSE)
}
\arguments{
\item{europe_countries_4}{If TRUE, returns values for 'GB-G', 'GB-N', 'DE-W', 'DE-E'.}
}
\description{
Return a color for each European country from the SR palette.
}
\examples{
palette_europe_countries <- palette_europe_countries( europe_countries_4 = TRUE)
}
\keyword{palette}
