% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_palette.R
\name{reprex_palette}
\alias{reprex_palette}
\title{Return the Reprex palette}
\usage{
reprex_palette(ext = FALSE)
}
\arguments{
\item{ext}{Defaults to FALSE. If true, adds five colors}
}
\value{
A named character vector of colors in a discrete palette.
}
\description{
Endre Koronczi Palette (c) 2015
}
\details{
A palette of \code{#007CBB}, \code{#4EC0E4}, \code{#00843A},
\code{#3EA135}, \code{#DB001C}, \code{#5C2320}, \code{#4E115A}, \code{#00348A},
\code{#BAC615}, \code{#FAE000}, \code{#E88500}, \code{#E4007F}.
}
\examples{
reprex_palette(ext = TRUE)
}
\seealso{
Other color functions: 
\code{\link{add_greys}()}
}
\concept{color functions}
