% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_colors.R
\name{create_palette}
\alias{create_palette}
\title{Create colors}
\usage{
create_palette(color1, color2, palette_length = 5, used_colors = NULL)
}
\arguments{
\item{palette_length}{How many shades to add, including the two starting colors (white and \code{x}).}

\item{x}{The color to start with.}
}
\description{
Create colors
}
