% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dir_files.R
\name{rename_dir_files}
\alias{rename_dir_files}
\title{Rename and copy image files from one directory to another}
\usage{
rename_dir_files(path_to_files, new_path, common_name)
}
\arguments{
\item{path_to_files}{Source directory containing image files}

\item{new_path}{Destination directory}

\item{common_name}{Base name for the renamed files}
}
\description{
Rename and copy image files from one directory to another
}
