% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dir_files.R
\name{rename_dir_files}
\alias{rename_dir_files}
\title{Rename files in directory}
\usage{
rename_dir_files(path_to_files, new_path, common_name)
}
\arguments{
\item{path_to_files}{Path to the files to rename}

\item{new_path}{New path}

\item{common_name}{Common file name part as a string.}
}
\value{
The renamed files in \code{new_path}
}
\description{
Rename files in directory
}
